# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: NCKH SaaS Multi-tenancy API
  version: 0.1.0
  description: >-
    Hợp đồng REST cho control plane và application plane. Các API nghiệp vụ lấy
    tenant từ host và JWT; không nhận tenant_id từ payload để phân quyền.
servers:
  - url: http://accounts.localhost:8080/api/v1
    description: Control host local
  - url: http://{tenant}.localhost:8080/api/v1
    description: Tenant host local
    variables:
      tenant:
        default: pool-demo
security:
  - bearerAuth: []
...
paths:
...
  /tenants:
    get:
      tags: [Tenants]
      operationId: listMyTenants
      responses:
        '200':
          description: Active memberships of current user
          content:
            application/json:
              schema:
                type: array
                items: {$ref: '#/components/schemas/TenantView'}
        default: {$ref: '#/components/responses/Error'}
    post:
      tags: [Tenants]
      operationId: createTenant
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CreateTenantRequest'}
      responses:
        '200':
          description: Tenant in PENDING_PAYMENT state
          content:
            application/json:
              schema: {$ref: '#/components/schemas/TenantView'}
        default: {$ref: '#/components/responses/Error'}
...
  /members:
    get:
      tags: [Tenants]
      operationId: listTenantMembers
      responses:
        '200':
          description: Active tenant members
          content:
            application/json:
              schema:
                type: array
                items: {$ref: '#/components/schemas/MemberView'}
        default: {$ref: '#/components/responses/Error'}
  /members/invitations:
    get:
      tags: [Tenants]
      operationId: listTenantInvitations
      responses:
        '200':
          description: Tenant-scoped invitation status ordered newest first
          content:
            application/json:
              schema:
                type: array
                items: {$ref: '#/components/schemas/InvitationView'}
        default: {$ref: '#/components/responses/Error'}
    post:
      tags: [Tenants]
      operationId: createTenantInvitation
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/InviteMemberRequest'}
      responses:
        '200':
          description: Pending invitation and one-time local delivery token
          content:
            application/json:
              schema: {$ref: '#/components/schemas/InvitationCreatedView'}
        default: {$ref: '#/components/responses/Error'}
  /members/invitations/{invitationId}:
    delete:
      tags: [Tenants]
      operationId: revokeTenantInvitation
      parameters:
        - name: invitationId
          in: path
          required: true
          schema: {$ref: '#/components/schemas/UUID'}
      responses:
        '200': {description: Pending invitation revoked idempotently}
        default: {$ref: '#/components/responses/Error'}
  /members/{membershipId}:
    parameters:
      - $ref: '#/components/parameters/MembershipId'
    patch:
      tags: [Tenants]
      operationId: updateTenantMember
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/UpdateMemberRequest'}
      responses:
        '200':
          description: Updated membership and incremented security version when changed
          content:
            application/json:
              schema: {$ref: '#/components/schemas/MemberView'}
        default: {$ref: '#/components/responses/Error'}
    delete:
      tags: [Tenants]
      operationId: revokeTenantMember
      responses:
        '200': {description: Membership revoked}
        default: {$ref: '#/components/responses/Error'}
  /members/{membershipId}/role:
    patch:
      tags: [Tenants]
      operationId: changeTenantMemberRole
      parameters:
        - $ref: '#/components/parameters/MembershipId'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ChangeRoleRequest'}
      responses:
        '200':
          description: Updated membership
          content:
            application/json:
              schema: {$ref: '#/components/schemas/MemberView'}
        default: {$ref: '#/components/responses/Error'}
  /members/{membershipId}/transfer-ownership:
    post:
      tags: [Tenants]
      operationId: transferTenantOwnership
      parameters:
        - $ref: '#/components/parameters/MembershipId'
      responses:
        '200':
          description: Target is the sole Owner and the previous Owner is now Admin
          content:
            application/json:
              schema: {$ref: '#/components/schemas/MemberView'}
        default: {$ref: '#/components/responses/Error'}
  /invitations/{token}:
    get:
      tags: [Tenants]
      operationId: previewTenantInvitation
      security: []
      parameters:
        - name: token
          in: path
          required: true
          schema: {type: string, minLength: 32, maxLength: 200}
      responses:
        '200':
          description: Invitation preview addressed by an opaque capability token
          content:
            application/json:
              schema: {$ref: '#/components/schemas/InvitationView'}
        default: {$ref: '#/components/responses/Error'}
  /invitations/{token}/accept:
    post:
      tags: [Tenants]
      operationId: acceptTenantInvitation
      parameters:
        - name: token
          in: path
          required: true
          schema: {type: string, minLength: 32, maxLength: 200}
      responses:
        '200':
          description: Invitation accepted idempotently by the matching account
          content:
            application/json:
              schema: {$ref: '#/components/schemas/InvitationView'}
        default: {$ref: '#/components/responses/Error'}
  /invitations/{token}/reject:
    post:
      tags: [Tenants]
      operationId: rejectTenantInvitation
      parameters:
        - name: token
          in: path
          required: true
          schema: {type: string, minLength: 32, maxLength: 200}
      responses:
        '200':
          description: Invitation rejected idempotently by the matching account
          content:
            application/json:
              schema: {$ref: '#/components/schemas/InvitationView'}
        default: {$ref: '#/components/responses/Error'}
...
  /resources:
    get:
      tags: [Resources]
      operationId: listResources
      responses:
        '200':
          description: Resources uploaded by or project-visible to current user
          content:
            application/json:
              schema:
                type: array
                items: {$ref: '#/components/schemas/ResourceView'}
        default: {$ref: '#/components/responses/Error'}
    post:
      tags: [Resources]
      operationId: uploadResource
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: {type: string, format: binary}
      responses:
        '200':
          description: Resource stored in tenant namespace
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ResourceView'}
        default: {$ref: '#/components/responses/Error'}
  /resources/links:
    post:
      tags: [Resources]
      operationId: createLinkResource
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CreateLinkResourceRequest'}
      responses:
        '200':
          description: Validated HTTP(S) link created in the tenant resource library
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ResourceView'}
        default: {$ref: '#/components/responses/Error'}
  /resources/quota:
    get:
      tags: [Resources]
      operationId: resourceQuota
      responses:
        '200':
          description: Aggregate quota for tenant administrators
          content:
            application/json:
              schema: {$ref: '#/components/schemas/QuotaView'}
        default: {$ref: '#/components/responses/Error'}
  /resources/{resourceId}/download-url:
    get:
      tags: [Resources]
      operationId: createResourceDownloadUrl
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          description: Short-lived authorized URL
          content:
            application/json:
              schema: {$ref: '#/components/schemas/DownloadUrl'}
        default: {$ref: '#/components/responses/Error'}
  /resources/{resourceId}/tasks/{taskId}:
    post:
      tags: [Resources]
      operationId: attachResourceToTask
      parameters:
        - $ref: '#/components/parameters/ResourceId'
        - $ref: '#/components/parameters/TaskId'
      responses:
        '200': {description: Resource attached idempotently}
        default: {$ref: '#/components/responses/Error'}
    delete:
      tags: [Resources]
      operationId: detachResourceFromTask
      parameters:
        - $ref: '#/components/parameters/ResourceId'
        - $ref: '#/components/parameters/TaskId'
      responses:
        '200': {description: Resource detached from task}
        default: {$ref: '#/components/responses/Error'}
  /resources/{resourceId}:
    delete:
      tags: [Resources]
      operationId: deleteResource
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200': {description: Resource metadata soft-deleted; file cleanup is handled by outbox}
        default: {$ref: '#/components/responses/Error'}
...
  /admin/tenants:
    get:
      tags: [Administration]
      operationId: listAdminTenants
      parameters:
        - name: page
          in: query
          schema: {type: integer, minimum: 0, default: 0}
        - name: size
          in: query
          schema: {type: integer, minimum: 1, maximum: 100, default: 20}
        - name: search
          in: query
          schema: {type: string, default: ''}
        - name: status
          in: query
          schema: {$ref: '#/components/schemas/TenantStatus'}
        - name: placement
          in: query
          schema: {$ref: '#/components/schemas/TenantPlacement'}
      responses:
        '200':
          description: Control-plane tenant page for system administrators
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AdminTenantPage'}
        default: {$ref: '#/components/responses/Error'}
  /admin/tenants/{tenantId}:
    get:
      tags: [Administration]
      operationId: getAdminTenant
      parameters:
        - $ref: '#/components/parameters/TenantId'
      responses:
        '200':
          description: Control-plane tenant detail, latest payment and provisioning transition history
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AdminTenantDetailView'}
        default: {$ref: '#/components/responses/Error'}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    TenantId:
      name: tenantId
      in: path
      required: true
      schema: {$ref: '#/components/schemas/UUID'}
    MembershipId:
      name: membershipId
      in: path
      required: true
      schema: {$ref: '#/components/schemas/UUID'}
...
    TaskId:
      name: taskId
      in: path
      required: true
      schema: {$ref: '#/components/schemas/UUID'}
    ResourceId:
      name: resourceId
      in: path
      required: true
      schema: {$ref: '#/components/schemas/UUID'}
...
  responses:
    Error:
      description: Structured API error
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ApiError'}
  schemas:
    UUID: {type: string, format: uuid}
    Instant: {type: string, format: date-time}
    TenantRole: {type: string, enum: [OWNER, ADMIN, MEMBER]}
    InvitationStatus: {type: string, enum: [PENDING, ACCEPTED, REJECTED, REVOKED, EXPIRED]}
...
    TenantTier: {type: string, enum: [STARTER, PROFESSIONAL, ENTERPRISE]}
    TenantPlacement: {type: string, enum: [POOL, SILO_DATABASE]}
    TenantStatus: {type: string, enum: [PENDING_PAYMENT, PROVISIONING, ACTIVE, FAILED, SUSPENDED]}
    PaymentStatus: {type: string, enum: [CREATED, PENDING, SUCCEEDED, FAILED, EXPIRED]}
    ProvisioningStatus: {type: string, enum: [QUEUED, RUNNING, SUCCEEDED, RETRYABLE_FAILED, FAILED_ROLLED_BACK, ROLLBACK_FAILED]}
...
    ApiError:
      type: object
      required: [timestamp, status, code, message, path, fieldErrors]
      properties:
        timestamp: {$ref: '#/components/schemas/Instant'}
        status: {type: integer}
        code: {type: string}
        message: {type: string}
        path: {type: string}
        fieldErrors: {type: object, additionalProperties: {type: string}}
        requestId: {type: [string, 'null']}
...
    TenantView:
      type: object
      required: [id, slug, name, tier, status, placement, role]
      properties:
        id: {$ref: '#/components/schemas/UUID'}
        slug: {type: string}
        name: {type: string}
        tier: {$ref: '#/components/schemas/TenantTier'}
        status: {$ref: '#/components/schemas/TenantStatus'}
        placement: {$ref: '#/components/schemas/TenantPlacement'}
        role: {$ref: '#/components/schemas/TenantRole'}
...
    CreateTenantRequest:
      type: object
      required: [slug, name, tier, placement]
      properties:
        slug: {type: string, pattern: '^[a-z0-9](?:[a-z0-9-]{0,61}[a-z0-9])?$'}
        name: {type: string}
        tier: {$ref: '#/components/schemas/TenantTier'}
        placement: {$ref: '#/components/schemas/TenantPlacement'}
    MemberView:
      type: object
      required: [membershipId, userId, email, displayName, role, active, securityVersion]
      properties:
        membershipId: {$ref: '#/components/schemas/UUID'}
        userId: {$ref: '#/components/schemas/UUID'}
        email: {type: string, format: email}
        displayName: {type: string}
        role: {$ref: '#/components/schemas/TenantRole'}
        active: {type: boolean}
        securityVersion: {type: integer, format: int64}
    InviteMemberRequest:
      type: object
      required: [email, role]
      properties:
        email: {type: string, format: email}
        role: {type: string, enum: [ADMIN, MEMBER]}
    InvitationView:
      type: object
      required: [id, tenantId, tenantSlug, tenantName, email, role, status, expiresAt]
      properties:
        id: {$ref: '#/components/schemas/UUID'}
        tenantId: {$ref: '#/components/schemas/UUID'}
        tenantSlug: {type: string}
        tenantName: {type: string}
        email: {type: string, format: email}
        role: {type: string, enum: [ADMIN, MEMBER]}
        status: {$ref: '#/components/schemas/InvitationStatus'}
        expiresAt: {$ref: '#/components/schemas/Instant'}
        respondedAt: {anyOf: [{$ref: '#/components/schemas/Instant'}, {type: 'null'}]}
    InvitationCreatedView:
      type: object
      required: [invitation, token, acceptancePath]
      properties:
        invitation: {$ref: '#/components/schemas/InvitationView'}
        token: {type: string, minLength: 32, writeOnly: true}
        acceptancePath: {type: string}
    UpdateMemberRequest:
      type: object
      required: [role, active]
      properties:
        role: {$ref: '#/components/schemas/TenantRole'}
        active: {type: boolean}
    ChangeRoleRequest:
      type: object
      required: [role]
      properties: {role: {$ref: '#/components/schemas/TenantRole'}}
...
    ResourceView:
      type: object
      required: [id, originalName, contentType, sizeBytes, uploadedBy, createdAt, taskCount, kind, taskIds]
      properties:
        id: {$ref: '#/components/schemas/UUID'}
        originalName: {type: string}
        storageKey:
          type: [string, 'null']
          description: Server-generated FILE key; clients must not use it for authorization
        contentType: {type: string}
        sizeBytes: {type: integer, format: int64}
        uploadedBy: {$ref: '#/components/schemas/UUID'}
        createdAt: {$ref: '#/components/schemas/Instant'}
        taskCount: {type: integer, format: int64}
        kind: {type: string, enum: [FILE, LINK]}
        linkUrl: {type: [string, 'null'], format: uri}
        taskIds:
          type: array
          items: {$ref: '#/components/schemas/UUID'}
    CreateLinkResourceRequest:
      type: object
      required: [name, url]
      properties:
        name: {type: string, maxLength: 255}
        url: {type: string, format: uri, maxLength: 2000}
    QuotaView:
      type: object
      required: [usedBytes, limitBytes]
      properties:
        usedBytes: {type: integer, format: int64}
        limitBytes: {type: integer, format: int64}
    DownloadUrl:
      type: object
      required: [url, expiresAt]
      properties:
        url: {type: string, format: uri-reference}
        expiresAt: {$ref: '#/components/schemas/Instant'}
...
    AdminTenantView:
      type: object
      required: [id, name, slug, tier, status, memberCount, createdAt]
      properties:
        id: {$ref: '#/components/schemas/UUID'}
        name: {type: string}
        slug: {type: string}
        tier: {$ref: '#/components/schemas/TenantTier'}
        placement: {anyOf: [{$ref: '#/components/schemas/TenantPlacement'}, {type: 'null'}]}
        status: {$ref: '#/components/schemas/TenantStatus'}
        provisioningStatus: {anyOf: [{$ref: '#/components/schemas/ProvisioningStatus'}, {type: 'null'}]}
        memberCount: {type: integer, format: int64}
        createdAt: {$ref: '#/components/schemas/Instant'}
        lastError: {type: [string, 'null']}
        paymentStatus: {anyOf: [{$ref: '#/components/schemas/PaymentStatus'}, {type: 'null'}]}
        paymentProvider: {type: [string, 'null']}
    AdminPaymentView:
      type: object
      required: [id, provider, status, amountMinor, currency, createdAt]
      properties:
        id: {$ref: '#/components/schemas/UUID'}
        provider: {type: string}
        status: {$ref: '#/components/schemas/PaymentStatus'}
        amountMinor: {type: integer, format: int64}
        currency: {type: string}
        createdAt: {$ref: '#/components/schemas/Instant'}
    AdminProvisioningView:
      type: object
      required: [id, status, attempts, createdAt]
      properties:
        id: {$ref: '#/components/schemas/UUID'}
        status: {$ref: '#/components/schemas/ProvisioningStatus'}
        attempts: {type: integer}
        lastErrorCode: {type: [string, 'null']}
        lastErrorMessage: {type: [string, 'null']}
        nextAttemptAt: {anyOf: [{$ref: '#/components/schemas/Instant'}, {type: 'null'}]}
        createdAt: {$ref: '#/components/schemas/Instant'}
    ProvisioningEventView:
      type: object
      required: [id, toStatus, attempt, createdAt]
      properties:
        id: {$ref: '#/components/schemas/UUID'}
        fromStatus: {anyOf: [{$ref: '#/components/schemas/ProvisioningStatus'}, {type: 'null'}]}
        toStatus: {$ref: '#/components/schemas/ProvisioningStatus'}
        attempt: {type: integer}
        errorCode: {type: [string, 'null']}
        message: {type: [string, 'null']}
        createdAt: {$ref: '#/components/schemas/Instant'}
    AdminTenantDetailView:
      type: object
      required: [tenant, events]
      properties:
        tenant: {$ref: '#/components/schemas/AdminTenantView'}
        payment: {anyOf: [{$ref: '#/components/schemas/AdminPaymentView'}, {type: 'null'}]}
        provisioning: {anyOf: [{$ref: '#/components/schemas/AdminProvisioningView'}, {type: 'null'}]}
        events:
          type: array
          items: {$ref: '#/components/schemas/ProvisioningEventView'}
    AdminTenantPage:
      type: object
      required: [items, page, size, totalItems, totalPages]
      properties:
        items: {type: array, items: {$ref: '#/components/schemas/AdminTenantView'}}
        page: {type: integer}
        size: {type: integer}
        totalItems: {type: integer, format: int64}
        totalPages: {type: integer}
...